requirements.txt
aiohttp==3.8.6
aiohttp-index==0.1
python-engineio==3.8.1
python-socketio==4.1.0
pandas==2.0.3
numpy==1.24.3
scipy==1.10.1
firebase-admin==6.2.0
websockets==8.1
requests==2.31.0